+++ common/src/main/java/com/alien/common/data/AlienAdvancements.java
@@ -9,6 +9,8 @@ public class AlienAdvancements {
 
     public static final BLibAdvancement KILL_A_HARBINGER = create("kill_a_harbinger");
 
+    public static final BLibAdvancement KILL_A_LINEAGE = create("kill_a_lineage");
+
     public static final BLibAdvancement KILL_A_ROYAL_ALIEN = create("kill_a_royal_alien");
 
     public static final BLibAdvancement KILL_ALL_ALIENS = create("kill_all_aliens");